+++ configs/import.js
@@ -1,4 +1,3 @@
-import importPlugin from "eslint-plugin-import";
 import simpleImportSort from "eslint-plugin-simple-import-sort";
 import unusedImports from "eslint-plugin-unused-imports";
 
@@ -6,7 +5,6 @@ export const importConfig = [
   {
     files: ["**/*.{js,ts}"], // For both JS and TS files
     plugins: {
-      import: importPlugin,
       "simple-import-sort": simpleImportSort,
       "unused-imports": unusedImports,
     },